+++ src/pptlive/__init__.py
@@ -28,7 +28,7 @@ from ._charts import Chart
 from ._edit import EditScope
 from ._presentation import Presentation, PresentationCollection
 from ._selection import SelectionInfo, SelectionSnapshot
-from ._shapes import PlaceholderShape, Shape, ShapeCollection
+from ._shapes import PlaceholderShape, Shape, ShapeById, ShapeCollection
 from ._show import SlideShow
 from ._slides import Slide, SlideCollection
 from ._smartart import SmartArt
@@ -73,6 +73,7 @@ __all__ = [
     "SelectionInfo",
     "SelectionSnapshot",
     "Shape",
+    "ShapeById",
     "ShapeCollection",
     "Slide",
     "SlideCollection",